# doc/generate/openapi.yml
openapi: "3.0.3"
servers:
  - url: http://localhost
    description: "Defined by server config / helm chart"
info:
  title: Acyl API
  description: "The Acyl API provides authenticated users options manage dynamic environments."
  contact:
    name: Benjamen Keroack
    url: "https://github.com/Pluto-tv/acyl"
    email: benjamen@dollarshaveclub.com
  license:
    name: MIT
    url: "https://opensource.org/licenses/MIT"
  version: v0.9.0
...
security:
  - api_key: []
components:
  securitySchemes:
    api_key:
      description: "API Key or aborts with Unauthorized"
      type: apiKey
      name: API-Key
      in: header
  schemas:
    QAEnvironmentEvent:
      type: object
      description: "This represents the default environment"
      properties:
        timestamp:
          type: string
          format: date-time
        message:
          type: string
    RefMap:
      type: object
      description: "This represents a map of repo name to git reference (branch)"
      properties:
        string:
          type: string
    CommitShaMap:
      type: object
      description: "This represents a map of repo name to full commit shas"
      properties:
        string:
          type: string
    QAEnvironment_v2:
      type: object
      description: "This represents the versioned (v2) environment that includes prior version changes, and adds source_ref field"
      properties:
        name:
          type: string
        created:
          type: string
          format: date-time
        events:
          $ref: '#/components/schemas/QAEnvironmentEvent'
        hostname:
          type: string
        qa_type:
          deprecated: true
          type: string
          description: "**DEPRECATED** not supported with Nitro v2 or newer"
        user:
          type: string
          description: "GitHub username of the user who opened the triggering PR."
        repo:
          type: string
          description: "GitHub repository, including organization if applicable (e.g. Pluto-tv/acyl)"
        pull_request:
          type: integer
          format: uint
          description: "GitHub Pull Request value"
        source_sha:
          type: string
          description: "GitHub SHA for the dev branch"
        base_sha:
          type: string
          description: "GitHub HEAD commit SHA for the branch the PR will merge into (eg, master)."
        source_branch:
          type: string
          description: "GitHub name of the dev branch"
        base_branch:
          type: string
          description: "GitHub branch name the PR will merge into (eg, master)."
        raw_status:
          type: string
        status:
          type: integer
          description: "Integer based on EnvironmentStatus; UnknownStatus: 0, Spawned: 1, Success: 2, Failure: 3, Destroyed: 4, Updating: 5, Cancelled: 6"
        ref_map:
          $ref: '#/components/schemas/RefMap'
        amino_service_to_port:
          deprecated: true
          type: object
          description: "**DEPRECATED** not supported with Nitro v2 or newer"
          properties:
            string:
              type: integer
              format: int64
        amino_kubernetes_namespace:
          deprecated: true
          type: string
          description: "**DEPRECATED** not supported with Nitro v2 or newer"
        amino_environment_id:
          deprecated: true
          type: integer
          description: "**DEPRECATED** not supported with Nitro v2 or newer"
        commit_sha_map:
          $ref: '#/components/schemas/CommitShaMap'
        source_ref:
          type: string
    EventLog:
      description: "This represents the event log"
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "Acyl event log id (uuid)"
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        env_name:
          type: string
          description: "Random generated Acyl environment name"
        repo:
          type: string
          description: "GitHub repository, including organization if applicable (e.g. Pluto-tv/acyl)"
        pull_request:
          type: integer
          format: uint
          description: "GitHub Pull Request value"
        webhook_payload:
          description: "Value of the `payload` field returned in the webhook response"
          type: string
          format: byte
        log:
          description: "A time-ordered array of debug log messages associated with this specific event for the environment"
          type: string
          format: array
  parameters:
    readOnlyAPIKey:
      name: "API Key"
      in: header
      description: "Read Only API Key required for this endpoint"
      required: true
      schema:
        type: string
        format: uuid
...
    prParam:
      name: pr
      in: query
      description: "GitHub Pull Request value, required if repo parameter included"
      required: false
      schema:
        type: string
        format: integer
    repoParam:
      name: repo
      in: query
      description: "GitHub repository, including organization if applicable (e.g. Pluto-tv/acyl)"
      required: false
      schema:
        type: string
    statusParam:
      name: status
      in: query
      description: "Integer based on EnvironmentStatus; UnknownStatus: 0, Spawned: 1, Success: 2, Failure: 3, Destroyed: 4, Updating: 5, Cancelled: 6"
      required: false
      schema:
        type: string
        format: integer
    sourceBranchParam:
      name: source_branch
      in: query
      description: "GitHub name of the dev branch"
      required: false
      schema:
        type: string
    sourceShaParam:
      name: source_sha
      in: query
      description: "GitHub SHA for the dev branch"
      required: false
      schema:
        type: string
    userParam:
      name: user
      in: query
      description: "GitHub username of the user who opened the triggering PR."
      required: false
      schema:
        type: string
    envNameParam:
      name: name
      in: path
      description: "Acyl environment name (e.g. foo-bar-baz)"
      required: true
      schema:
        type: string
    eventLogIdParam:
      name: id
      in: path
      description: "Acyl event log id (uuid)"
      required: true
      schema:
        type: string
        format: uuid
  responses:
    400:
      description: "Bad Request"
    404:
      description: "Not Found"
    500:
      description: "Internal Server Error"
  examples:
...
    env:
      summary: "Environment Details Response Example"
      value:
        name: eparchial-manifest-destiny
        created: '2021-03-24T23:09:06.832246Z'
        events:
          - created: '2021-03-24T23:08:46.600387605Z'
            message: spawned
          - created: '2021-03-24T23:08:46.684560586Z'
            message: 'webhook event id: ecd604a8-cf63-4536-8bdf-c83f4dd27d50'
          - created: '2021-03-24T23:08:54.200958318Z'
            message: 'creating namespace: nitro-96517-eparchial-manifest-destiny'
          - created: '2021-03-24T23:08:54.263118013Z'
            message: 'building container: organization/reliable-backend:a6593cc6e2ff530ef0ccd872b9f4b6466598cead'
          - created: '2021-03-24T23:08:54.359604057Z'
            message: installing tiller into namespace
          - created: '2021-03-24T23:08:54.394465482Z'
            message: 'building container: organization/awesome-app:b519def177c7ca1c855d15539b5a3314eeb5ff90'
          - created: '2021-03-24T23:08:54.397195464Z'
            message: 'building container: organization/cool-app:0a63aa79565879e88f0eec36ed19a81a965d7479'
          - created: '2021-03-24T23:08:56.810781995Z'
            message: Marked as Cancelled
          - created: '2021-03-24T23:09:04.534629056Z'
            message: Marked as Failure
          - created: '2021-03-24T23:09:04.544497904Z'
            message: Marked as Failure
          - created: '2021-03-24T23:09:06.824599562Z'
            message: Marked as Spawned
          - created: '2021-03-24T23:09:06.827189537Z'
            message: reusing environment record for webhook event c643f8eb-e903-4155-9cf5-505f1923beb0
          - created: '2021-03-24T23:09:06.844013853Z'
            message: 'webhook event id: c643f8eb-e903-4155-9cf5-505f1923beb0'
          - created: '2021-03-24T23:09:14.400262282Z'
            message: 'creating namespace: nitro-87012-eparchial-manifest-destiny'
          - created: '2021-03-24T23:09:14.445484499Z'
            message: 'building container: organization/awesome-app:610f09cefcdec3e3468d14a781037d9c58a7b95f'
          - created: '2021-03-24T23:09:14.497301671Z'
            message: 'building container: organization/cool-app:0a63aa79565879e88f0eec36ed19a81a965d7479'
          - created: '2021-03-24T23:09:14.499761749Z'
            message: 'building container: organization/reliable-backend:a6593cc6e2ff530ef0ccd872b9f4b6466598cead'
          - created: '2021-03-24T23:09:14.518606837Z'
            message: installing tiller into namespace
          - created: '2021-03-24T23:09:17.912562605Z'
            message: 'build finished: organization/cool-app: b688e242-f65f-4a79-ad35-a7b5ee8ed9dd'
          - created: '2021-03-24T23:09:18.171093466Z'
            message: 'build finished: organization/reliable-backend: 84e52bbf-bdd2-40f4-b172-4d3cdfddcc5a'
          - created: '2021-03-24T23:09:24.602140295Z'
            message: image build complete; installing chart for reliable-backend
          - created: '2021-03-24T23:09:24.603825869Z'
            message: image build complete; installing chart for cool-app
          - created: '2021-03-24T23:09:46.501416625Z'
            message: image build still pending; waiting to install chart for organization-awesome-app
          - created: '2021-03-24T23:09:56.506436537Z'
            message: image build still pending; waiting to install chart for organization-awesome-app
          - created: '2021-03-24T23:10:06.511985081Z'
            message: image build still pending; waiting to install chart for organization-awesome-app
          - created: '2021-03-24T23:10:16.518812277Z'
            message: image build still pending; waiting to install chart for organization-awesome-app
          - created: '2021-03-24T23:14:23.755408995Z'
            message: 'build finished: organization/awesome-app: 0ffd2b19-9844-4c89-baba-b51d6a6c338a'
          - created: '2021-03-24T23:14:26.674776938Z'
            message: image build complete; installing chart for organization-awesome-app
          - created: '2021-03-24T23:16:15.291184822Z'
            message: 'all charts installed; release names: map[cool-app:cool-app reliable-backend:reliable-backend
                            organization-awesome-app:organization-awesome-app]'
          - created: '2021-03-24T23:16:15.30458093Z'
            message: Marked as Success
        hostname: ''
        qa_type: ''
        user: bobsmith
        repo: organization/awesome-app
        pull_request: 337
        source_sha: 610f09cefcdec3e3468d14a781037d9c58a7b95f
        base_sha: f7e8e09ea634fdedb37623afc0f7835b7d0ad417
        source_branch: awesome-feature
        base_branch: master
        source_ref: awesome-feature
        status: Success
        ref_map:
          organization/cool-app: master
          organization/reliable-backend: master
          organization/awesome-app: awesome-feature
        commit_sha_map:
          organization/cool-app: 0a63aa79565879e88f0eec36ed19a81a965d7479
          organization/reliable-backend: a6593cc6e2ff530ef0ccd872b9f4b6466598cead
          organization/awesome-app: 610f09cefcdec3e3468d14a781037d9c58a7b95f
        amino_service_to_port: {}
        amino_kubernetes_namespace: ''
        amino_environment_id: 0
...
paths:
  /v2/envs/_search:
    get:
      tags:
        - v2
      summary: "Search environments from optional parameters, minimum one parameter required; multiple parameters are combined with a logical AND: (PR = n AND repo = foo/bar AND status = x)"
      operationId: "# Search Environments"
      parameters:
        - $ref: '#/components/parameters/readOnlyAPIKey'
        - $ref: '#/components/parameters/prParam'
        - $ref: '#/components/parameters/repoParam'
        - $ref: '#/components/parameters/statusParam'
        - $ref: '#/components/parameters/sourceBranchParam'
        - $ref: '#/components/parameters/sourceShaParam'
        - $ref: '#/components/parameters/userParam'
      responses:
        200:
          description: "Returns list of detailed environments as defined by QAEnvironment (v2)"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QAEnvironment_v2'
              example:
                $ref: '#/components/examples/envs'
        400:
          $ref: '#/components/responses/400'
        500:
          $ref: '#/components/responses/500'
  /v2/envs/{name}:
    get:
      tags:
        - v2
      summary: "Get environment details"
      operationId: "# Environment Details"
      parameters:
        - $ref: '#/components/parameters/readOnlyAPIKey'
        - $ref: '#/components/parameters/envNameParam'
      responses:
        200:
          description: "Returns detailed environment as defined by QAEnvironment (v2)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QAEnvironment_v2'
              example:
                $ref: '#/components/examples/env'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'
  /v2/eventlog/{id}:
    get:
      tags:
        - v2
      summary: "Get event logs by id"
      operationId: "# Event Logs"
      parameters:
        - $ref: '#/components/parameters/readOnlyAPIKey'
        - $ref: '#/components/parameters/eventLogIdParam'
      responses:
        200:
          description: "Returns the EventLog that include debug logs for requested id, especially helpful troubleshooting failed environments"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventLog'
              example:
                $ref: '#/components/examples/eventLog'
        400:
          $ref: '#/components/responses/400'
        404:
          $ref: '#/components/responses/404'
        500:
          $ref: '#/components/responses/500'